Using `spawn_blocking`
